# api/openapi.yaml
openapi: 3.1.0
info:
  title: CarMarket API
  version: 0.1.0
  description: India used car marketplace REST API
servers:
  - url: http://localhost:8000/api/v1
    description: Local development
paths:
...
  /dealer-stores:
    post:
      summary: Create dealer store (upgrades user to dealer)
      tags: [dealer-stores]
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Created
  /dealer-stores/{slug}:
    get:
      summary: Public dealer store page
      tags: [dealer-stores]
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Store detail
  /dealer-stores/me:
    patch:
      summary: Update own dealer store
      tags: [dealer-stores]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Updated
  /dealer-stores/me/listings:
    get:
      summary: Dealer store inventory
      tags: [dealer-stores]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Listings
...
  /listings/{listing_id}/inquiries:
    post:
      summary: Send inquiry on listing
      tags: [inquiries]
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Inquiry created
  /inquiries/inbox:
    get:
      summary: Seller inquiry inbox
      tags: [inquiries]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Inbox
  /inquiries/sent:
    get:
      summary: Buyer sent inquiries
      tags: [inquiries]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Sent inquiries
  /inquiries/{inquiry_id}/accept:
    patch:
      summary: Accept inquiry (reveals seller phone to buyer)
      tags: [inquiries]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Accepted
  /inquiries/{inquiry_id}/decline:
    patch:
      summary: Decline inquiry
      tags: [inquiries]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Declined
...
  /favorites:
    get:
      summary: List favorite listings
      tags: [favorites]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Favorites
  /favorites/{listing_id}:
    post:
      summary: Add favorite
      tags: [favorites]
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Added
    delete:
      summary: Remove favorite
      tags: [favorites]
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Removed
...
  /admin/stats:
    get:
      summary: Platform statistics
      tags: [admin]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Stats
  /admin/users/{user_id}/role:
    patch:
      summary: Update user role
      tags: [admin]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Updated
  /admin/dealer-stores/{store_id}/verify:
    post:
      summary: Verify or reject dealer store
      tags: [admin]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Verified
  /admin/audit-logs:
    get:
      summary: List audit logs
      tags: [admin]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Audit logs
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT